增加DECLEAR_MSG_MAP_EX宏,以便将消息映射表可以写到cpp文件中。

+++ trunk/SOUI/src/control/SListView.cpp
@@ -650,7 +650,7 @@ namespace SOUI
         }
         m_lstItems.RemoveAll();
 
-        for(int i=0;i<m_itemRecycle.GetCount();i++)
+        for(int i=0;i<(int)m_itemRecycle.GetCount();i++)
         {
             SList<SItemPanel*> *pLstTypeItems = m_itemRecycle[i];
             SPOSITION pos = pLstTypeItems->GetHeadPosition();
